+++ packages/cds-odata-annotation-converter/test/setup.ts
@@ -3,7 +3,7 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import type { Diagnostic, Element } from '@sap-ux/odata-annotation-core';
 import type { AnnotationGroup, Annotation } from '@sap-ux/cds-annotation-parser';
-import { deserialize } from './deserialize-ast';
+import { deserialize } from './deserialize-ast.js';
 import cdsCompilerFacade from '@sap/ux-cds-compiler-facade';
 import type { CdsCompilerFacade } from '@sap/ux-cds-compiler-facade';