+++ packages/fiori-generator-shared/test/app-helpers/app-helper.test.ts
@@ -1,4 +1,4 @@
-import { getFlpId, getSemanticObject } from '../../src/app-helpers/app-helpers';
+import { getFlpId, getSemanticObject } from '../../src/app-helpers/app-helpers.js';
 
 describe('app-helper tests', () => {
     test('should return flp id', () => {