+++ packages/project-access/test/project/script.test.ts
@@ -2,7 +2,7 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { create as createStorage } from 'mem-fs';
 import { create } from 'mem-fs-editor';
-import { FileName, updatePackageScript, hasUI5CliV3 } from '../../src';
+import { FileName, updatePackageScript, hasUI5CliV3 } from '../../src/index.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));