+++ packages/btp-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/btp-utils
 
+## 2.0.1
+
+### Patch Changes
+
+- 21a3de7: FIX: TypeScript type errors in test files (ESM migration follow-up)
+
 ## 2.0.0
 
 ### Major Changes