+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/i18n
 
+## 1.0.1
+
+### Patch Changes
+
+- 21a3de7: FIX: TypeScript type errors in test files (ESM migration follow-up)
+- Updated dependencies [21a3de7]
+    - @sap-ux/text-document-utils@1.0.1
+
 ## 1.0.0
 
 ### Major Changes