+++ packages/ui-prompting/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/ui-prompting
 
+## 1.0.2
+
+### Patch Changes
+
+- 21a3de7: FIX: TypeScript type errors in test files (ESM migration follow-up)
+- Updated dependencies [21a3de7]
+    - @sap-ux/ui-components@3.0.2
+
 ## 1.0.1
 
 ### Patch Changes